==> 01_convo_chains.ipynb <==
# This notebook groups tweets into conversations and groups the
# conversations by company and saves this info in "data/comp_cll_df.csv"

# You need to first download twcs.zip (the raw data file containing
# all the tweets) from 
# https://www.kaggle.com/thoughtvector/customer-support-on-twitter/downloads/twcs.zip/10.
# You may have to create a free Kaggle account to do this.
# Extract twcs.csv to the data/ folder.

## importing modules

%matplotlib notebook
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import os
from shutil import rmtree
from glob import glob
from fnmatch import fnmatch
from dill import load, dump
from pprint import pprint


## read in data

df_raw = pd.read_csv('data/twcs.csv', nrows=1000)
print(df_raw.info())
df_raw.head()

## clean data a bit and SAVE

# convert 'created_at' to datetime format
df = df_raw[['created_at', 'author_id', 'text', 'tweet_id', 'response_tweet_id', 
         'in_response_to_tweet_id', 'inbound']].copy()
df['time'] = pd.to_datetime(df['created_at'], infer_datetime_format=True)

# rearrange columns and sort by time
df = df[['time', 'author_id', 'text', 'tweet_id', 'response_tweet_id', 
         'in_response_to_tweet_id', 'inbound']]
df.sort_values(by = "time", inplace = True)

# set tweet_id as index
df.set_index('tweet_id', inplace = True)

# standardize whitespace to avoid problems with saving as csv
df["text"] = df["text"].str.replace(r"\s+", " ", regex = True)

# save
df.to_csv("data/twcs_clean.csv")

### ENTRY POINT. load clean csv

df = pd.read_csv('data/twcs_clean.csv').set_index("tweet_id")
df["time"] = pd.to_datetime(df["time"], infer_datetime_format=True)
print(df.info())
df.head()

## make tweet chains for individual conversations

# save start_locs of convos
# a convo starts when 'in_response_to' is nan
def is_start_of_convo(rowtuple):
    
    # having a finite "in_response_to_tweet_id" 
    # means tweet is not convo starter
    if np.isfinite(
        getattr(rowtuple, "in_response_to_tweet_id")):
        return False
    
    else:
        return True
    
def get_start_locs(df = df):
    
    print("Computing convo_start_locs ...")
    
    # get the starter tups
    starter_tups_map = filter(is_start_of_convo,
                             df.itertuples())
    
    # get the locs of the starter tups
    tup_to_loc = lambda tup: tup.Index
    start_locs = list(map(tup_to_loc, starter_tups_map))
    
    return start_locs
    
convo_start_locs = get_start_locs()
        
print('Total convos =', len(convo_start_locs))
# print('convo_start_locs =', convo_start_locs)

# get the locs (and company) for a conversation given a start_loc
def get_convo_locs_n_comp(convo_start_loc, df = df):
    
    def get_next_locs(cur_locs, df = df):
        # recursively gets next locs
        
        next_locs = []
        for cur_loc in cur_locs:
            
            if cur_loc in df.index:
                resp_tw_ids_entry = df.loc[cur_loc]['response_tweet_id']
                
                if isinstance(resp_tw_ids_entry, str):
                    # next line splits up something like "45,65"
                    # in the response_tweet_id into a list of locs: [45, 65]
                    next_locs += list(map(int, resp_tw_ids_entry.split(",")))
                    
        if len(next_locs) > 0:
            next_locs += get_next_locs(next_locs)
        
        return next_locs
    
    convo_locs = [convo_start_loc] + get_next_locs([convo_start_loc])
    
    # take out non-existent locs
    convo_locs = list(filter(
        lambda loc: loc in df.index,
        convo_locs
        ))
    
    # make df and sort by time
    convo_df = df.loc[convo_locs]
    convo_df.sort_values(by = 'time', inplace = True)
    convo_locs = list(convo_df.index)
    
    # get company
    company = np.nan
    for rowtup in convo_df.itertuples():

        try:
            # if int, keep hunting
            _ = int(getattr(rowtup, "author_id"))
            
        except:
            # if not int, save
            company = getattr(rowtup, "author_id")
            break
    
    return convo_locs, company

start_loc = 18
convo_locs, company = get_convo_locs_n_comp(start_loc)
print("\nExample convo for \"" + company + "\":")
# print(df.loc[convo_locs][["author_id", "time", "text"]])

# temporary
print(df.loc[convo_locs][["text"]].values)

## SLOW! get convo_loc_lists for all start_locs

# cll: convo_loc_list
cll_comp_tups = list(map(get_convo_locs_n_comp, convo_start_locs))
pprint(cll_comp_tups[:5])
print(" ...")

### group convos by company

from collections import defaultdict

# making a dict where item is (company: clll)
# clll: cll_list i.e. list of conversations ("cll"s)

comp_clll_dict = defaultdict(list)

for cll, comp in cll_comp_tups:
    comp_clll_dict[comp].append(cll)
    
comp_clll_dict

### tool to save comp_clll_dict to csv

def save_locs_dict_to_csv(comp_clll_dict = comp_clll_dict):
    
    # [[1, 2], [3, 4, 5]] to "1,2 3,4,5"
    def list_list_to_str(ll):
        return " ".join(
            [",".join(list(map(str, l)))
                for l in ll]
            )
        
    clls_map = comp_clll_dict.values()
#     cll = list(comp_clll_dict.values())[0]
#     print()
    
    cll_strs = list(map(list_list_to_str, clls_map))
    
    # make df
    comp_cll_df = pd.DataFrame({
        "company": list(comp_clll_dict.keys()),
        "clll_str": cll_strs
        })
    
    comp_cll_df.to_csv("data/comp_cll_df.csv", index = False)
    
    return "data/comp_cll_df.csv"

print("comp_cll_df to csv:")
print(pd.read_csv(save_locs_dict_to_csv()).head())


### ENTRY POINT. tool to load comp_clll_dict from csv

def load_locs_dict_from_csv(path = "data/comp_cll_df.csv"):
    
    # "1,2 3,4,5" to [[1, 2], [3, 4, 5]]
    def str_to_ll(s):
        clstrs = s.split(" ")
        cll = [s.split(",") for s in clstrs]
        return cll
    
    # read csv
    df = pd.read_csv(path)
    comps = df["company"].values
    clll_strs = df["clll_str"]
    cllls_map = map(str_to_ll, clll_strs)
    
    # turn to dict
    comp_clll_dict = dict(zip(
        comps, cllls_map
        ))
    return comp_clll_dict

comp_clll_dict = load_locs_dict_from_csv()
print("comp_clll_dict_len:")
pprint(len(comp_clll_dict))
