==> 02_sentiment.ipynb <==
# get sentiments for all tweets and save to file

# for this notebook to work, you first need to set up a Google Cloud Natural 
# Language API project by following the steps under Prerequisites in 
# https://cloud.google.com/natural-language/docs/sentiment-tutorial

from gcp_sentiment import GCPSentiment
from dill import dump as dilldump

# function to get sentiment from GCP
credentialspath = "/home/vagrant/google_service_account_credentials/" +\
    "project_our-rock-218016.credentials.json"
get_sentence_sentiments = GCPSentiment(credentialspath
    ).get_sentence_sentiments

# sample use
get_sentence_sentiments([" ", "y'all gonna lie?. i'm tellin you go to smh"])

## read in tweets df

import pandas as pd

tweets_df = pd.read_csv("data/twcs_clean.csv", nrows=None) # small test
print(tweets_df.info())
tweets_df.head()



### saving sentiments as dill files, each containing
# a list of 1000 sentiments

import re, os
from time import sleep

tweets = tweets_df["text"].values

## strip handles e.g. @delta
handle_regobj = re.compile(r"@\w+")
def strip_handles(tweet,
    handle_regobj = handle_regobj):
    
    t_stripped = " ".join(handle_regobj.split(tweet))
    return t_stripped

print("   Stripping handles from tweets ...")
tweets = list(map(strip_handles, tweets))

# create save directory
if not os.path.exists("data/sentis/"): os.makedirs("data/sentis/")

batchsize = 1000
ix_end_batch = -1
while ix_end_batch != len(tweets) - 1:
    
    # creating batch
    ix_start_batch = ix_end_batch + 1
    ix_end_batch = min(len(tweets) - 1, ix_start_batch + batchsize - 1)
    batch_tweets = tweets[ix_start_batch : ix_end_batch + 1]
    
    # getting sentiments
    print("   Getting sentiments for " + str(ix_end_batch - ix_start_batch + 1) + 
          " tweets from " + str(ix_start_batch) + " to " + str(ix_end_batch) + 
          " ...")
    batch_sentis = get_sentence_sentiments(batch_tweets)
    sleep(10)
    print("   n_sentis =", len(batch_sentis))
    
    # saving to file
    path = "data/sentis/batch_sentis_" + str(ix_start_batch).zfill(9) + "_to_" + \
        str(ix_end_batch).zfill(9) + ".list.dill"
    
    with open(path, "wb") as f:
        dilldump(batch_sentis, f)
