==> 03_adding_sentiments_to_tweets_csv.ipynb <==
# put sentiments as a column in the tweets table

import pandas as pd

tweets_df = pd.read_csv("data/twcs_clean.csv")

import numpy as np
from dill import load as dillload
from glob import glob

# we'll read the sentiments into twcs_clean
# we begin by creating a np.array "sentis" of
# nans having length same as twcs_clean
sentis = np.full(shape = (len(tweets_df),), 
                 fill_value = np.nan)

## read in the sentiment dill files
sdill_paths = glob("data/sentis/*.list.dill")

# keep count of how many you're reading
n_sentis_read_total = 0
    
for sdill_path in sdill_paths[:]:
    
    # figure out start iloc
    pre_str = "_sentis_"
    post_str = "_to_"
    start_iloc = int(sdill_path[
        sdill_path.index(pre_str) + len(pre_str) :
        sdill_path.index(post_str)
        ])
    
    # figure out end iloc
    pre_str = "_to_"
    post_str = ".list"
    end_iloc = int(sdill_path[
        sdill_path.index(pre_str) + len(pre_str) :
        sdill_path.index(post_str)
        ])
    
    n_sentis_expected = end_iloc - start_iloc + 1
    
    sentilist = []
    with open(sdill_path, "rb") as f:
        sentilist = dillload(f)
    
    # check length for integrity
    n_sentis = len(sentilist)
    n_sentis_read_total += n_sentis
    if n_sentis != n_sentis_expected:
        print("   MISMATCH on " + sdill_path + ":")
        print("      n_sentis_expected", n_sentis_expected)
        print("      n_sentis", len(sentilist))
        
    else:
        sentis[start_iloc : end_iloc + 1] = sentilist
        
# integrity check
n_sentis_saved = np.sum(np.isfinite(sentis))
if n_sentis_read_total != n_sentis_saved:
    print("   MISMATCH on n_sentis:")
    print("      n_sentis_read_total", n_sentis_read_total)
    print("      n_sentis_saved", n_sentis_saved)
        
else:
    tweets_df["sentiment"] = sentis
    
tweets_df[["text", "sentiment"]].head()

# SLOW! save new tweets_df

tweets_df.to_csv("data/tweets_w_senti.csv", index = False)
