==> 06_averages_over_time.ipynb <==
# average satisfaction metrics over two periods:
# 2 weeks and 2 months

# read in and examine data
import pandas as pd

timeserie_df = pd.read_csv("data/timeseries.csv")

# fixing the index
timeserie_df = timeserie_df.set_index("time")
timeserie_df.index = pd.to_datetime(timeserie_df.index)

timeserie_df.head()

# function to average metric over some period

from scipy.stats import percentileofscore
from ranking import Ranking
import numpy as np

# keeping only relevant times
histstr2timedel_d = {"2 months": pd.Timedelta(days = 62), 
                     "2 weeks": pd.Timedelta(days = 14)}

# function to get ranks
def get_ranks(li):

    # get descending list
    li = np.array(li)
    desc_ixs = (-li).argsort()
    li_desc = li[desc_ixs]

    ## mapping from desc list to original list
    reg_ixs = range(len(li))
    # this will sort reg_ix by desc_ix
    desc_reg_ix_tups = sorted(zip(desc_ixs, reg_ixs))
    undo_desc_ixs = [t[1] for t in desc_reg_ix_tups]

    # ranking
    ranks = np.array(
        [tup[0] for tup in Ranking(li_desc)])[undo_desc_ixs] + 1
    
    return ranks

# function for averageness
def get_averageness(li):
    li = np.array(li)
    mean = li.mean()
    std = li.std()
    
    # set all to average
    avstrs = np.array(["average" for _ in li], dtype = "<U20")
    
    # find below / above averages
    avstrs[li < mean - std/2] = "below average"
    avstrs[li > mean + std/2] = "above average"
    
    return avstrs

def get_statmetdf(ts_df, metric, histstr,
    histstr2timedel_d = histstr2timedel_d):
    
    # keep only relevan period
    ts_df = ts_df.loc[
        ts_df.index[-1] - histstr2timedel_d[histstr] : None]

    # get the averages
    ts_met = ts_df[[c for c in list(ts_df) if 
                    metric in c
                    and "indavg" not in c]]
    met_avgs = ts_met.values.mean(axis = 0)
    met_avg_ser = pd.Series(met_avgs, index = list(ts_met))

    # creating df with raw metric vals
    statmetric_df = met_avg_ser.to_frame()
    statmetric_df = statmetric_df.rename(columns={0: metric})

    # fixing index
    statmetric_df = statmetric_df.sort_index()
    ixs = statmetric_df.index
    statmetric_df.index = [ix[None:ix.index("__")] for ix in ixs]
    statmetric_df.index.name = "company"

    # keeping metvals for reuse
    metvals = statmetric_df[metric].values

    # get rank
    ranks = get_ranks(metvals)
    statmetric_df[metric + "_rank"] = ranks
    
    # averageness
    statmetric_df[metric + "_averageness"] = get_averageness(metvals)
    
    # prepending all colnames with histstr
    hsprepend = "_".join(histstr.split())
    statmetric_df = statmetric_df.rename(columns = dict([
        (c, hsprepend + "_" + c) for c in list(statmetric_df)
        ]))
    
    return statmetric_df

get_statmetdf(timeserie_df, "cust_sent_improvement", "2 months")

# creating the statmet_df for both metrics and periods

# getting dfs for both metrics and both periods
metrics = ["issue_was_resolved", "cust_sent_improvement"]
histstrs = ["2 months", "2 weeks"]
statmet_dfs = []
for metric in metrics:
    for hs in histstrs:
        statmet_dfs.append(get_statmetdf(timeserie_df, metric, hs))
statmet_df = pd.concat(statmet_dfs, axis = "columns")

def get_combined_metrics(hspre, statmet_df = statmet_df):
    
    # getting percentiles
    resolved_scores = statmet_df[hspre + "_issue_was_resolved"]
    resolved_percs = resolved_scores.apply(
        lambda score: percentileofscore(resolved_scores, score))
    sentrise_scores = statmet_df[hspre + "_cust_sent_improvement"]
    sentrise_percs = sentrise_scores.apply(
        lambda score: percentileofscore(sentrise_scores, score))

    # weighted sum of percentiles
    percsums = (resolved_percs + sentrise_percs*2).round(2)

    # ranking percsums
    ranks = get_ranks(percsums)
    
    # getting averageness
    avstrs = get_averageness(percsums)
    
    return {"ranks": ranks, "avstrs": avstrs}

# assingning overall metrics
for histstr in histstrs:
    hspre = "_".join(histstr.split())
    combined_metrics = get_combined_metrics(hspre)
    statmet_df[hspre + "_rank"] = combined_metrics["ranks"]
    statmet_df[hspre + "_averageness"] = combined_metrics["avstrs"]
    
# saving
statmet_df.to_csv("data/static_metrics.csv")

# loading
statmet_df = pd.read_csv("data/static_metrics.csv").set_index("company")
statmet_df
